rename: api 파일명 접미사 제거후 통일
