+++ src/features/update-todo/api/index.ts
@@ -1,7 +1,1 @@
-// Todo API 세그먼트 Public API
-// export * from './todo-api';
-// export * from './queries';
-
 export * from './updateTodo';
-
-export * from './updateTodoApi';
